feature: add metric feature

+++ test/api.spec.ts
@@ -5,16 +5,41 @@ import 'mocha'
 
 import { fork } from 'child_process'
 
-describe('API', () => {
+describe('API', function () {
+  this.timeout(5000)
 
   describe('Notify', () => {
     it('should receive data', (done) => {
-      const child = fork(SpecUtils.buildTestPath('fixtures/apiChild.js'))
+      const child = fork(SpecUtils.buildTestPath('fixtures/apiNotifyChild.js'))
 
       child.on('message', msg => {
-        expect(msg).to.equal('myNotify')
-        child.kill('SIGINT')
-        done()
+        if (msg === 'myNotify') {
+          expect(msg).to.equal('myNotify')
+          child.kill('SIGINT')
+          done()
+        }
+      })
+    })
+  })
+
+  describe('Metrics', () => {
+    it('should receive data', (done) => {
+      const child = fork(SpecUtils.buildTestPath('fixtures/apiMetricsChild.js'))
+
+      child.on('message', res => {
+
+        if (res.type === 'axm:monitor') {
+          expect(res.data.hasOwnProperty('metricHistogram')).to.equal(true)
+          expect(res.data.metricHistogram.value).to.equal('10')
+          expect(res.data.metricHistogram.type).to.equal('metric/custom')
+          expect(res.data.hasOwnProperty('Loop delay')).to.equal(true)
+          expect(res.data.hasOwnProperty('Active handles')).to.equal(true)
+
+          if (res.data.hasOwnProperty('New space used size')) {
+            child.kill('SIGINT')
+            done()
+          }
+        }
       })
     })
   })
